feat(editor): add parameter dependency system and editor improvements
Major Features:
- Parameter dependency tracking for SubStateMachines with auto-linking
- Exit node support for nested state machine exits
- Any State -> Exit transitions for conditional exits
- Dependencies panel showing unresolved parameter links
- Parameter linking/aliasing between parent and child machines

Editor Improvements:
- Undo/redo support with view transform preservation
- Node deletion cleanup (dictionaries, edges, Any State transitions)
- Prevent deletion of special nodes (Any State, Exit)
- SubStateMachine inspector with dependency resolution UI
- Null guards and lazy initialization in inspectors

Performance Optimizations:
- Remove LINQ allocations across 40+ editor files
- Add GUIContentCache, IconCache, StringBuilderCache
- Add ListPool, HashSetPool, DictionaryPool, RectPool
- Cache popup options and state name arrays

Bug Fixes:
- Fix OnEnable null target issues in inspectors
- Fix parameter deletion not updating dependencies view
- Fix AnyStateExitTransition serialization (use HasAnyStateExitTransition)
- Fix GUI layout errors in AnyStateTransitionsInspector

+++ Runtime/Systems/UpdateStateMachineJob.cs
@@ -87,7 +87,7 @@ namespace DMotion
                 currentStateAnimationState,
                 ref stateMachineBlob,
                 parameters,
-                stateMachine.CurrentState.StateIndex,
+                (short)stateMachine.CurrentState.StateIndex,
                 out var transitionIndex);
 
             // If no Any State transition matched, check regular state transitions
